# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UAT Test Case Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /test-runs/{testRunId}/defects:
    get:
      operationId: listDefects
      tags: [defects]
      summary: List all defects for a test run
      parameters:
        - name: testRunId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of defects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Defect"
  /defects/{defectId}:
    get:
      operationId: getDefect
      tags: [defects]
      summary: Get a single defect with full detail
      parameters:
        - name: defectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Defect detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
  /defects/{defectId}/flag-bug:
    patch:
      operationId: flagDefectAsBug
      tags: [defects]
      summary: Flag a defect as a bug (Test Lead/Admin)
      parameters:
        - name: defectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Defect flagged as bug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
  /defects/{defectId}/flag-retest:
    patch:
      operationId: flagDefectForRetest
      tags: [defects]
      summary: Flag a defect for retesting
      parameters:
        - name: defectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagRetestBody"
      responses:
        "200":
          description: Defect flagged for retest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
  /defects/{defectId}/flag-accepted-by-business:
    patch:
      operationId: flagDefectAcceptedByBusiness
      tags: [defects]
      summary: Flag a defect as accepted by business (Test Lead/Admin)
      parameters:
        - name: defectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Defect accepted by business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
  /defects/{defectId}/business-accept:
    patch:
      operationId: businessAcceptDefect
      tags: [defects]
      summary: Business Owner accepts a defect
      parameters:
        - name: defectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessAcceptBody"
      responses:
        "200":
          description: Defect accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
  /defects/{defectId}/business-reject:
    patch:
      operationId: businessRejectDefect
      tags: [defects]
      summary: Business Owner rejects a defect
      parameters:
        - name: defectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessRejectBody"
      responses:
        "200":
          description: Defect rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
  /defects/{defectId}/notes:
    post:
      operationId: addDefectNote
      tags: [defects]
      summary: Add a note to a defect
      parameters:
        - name: defectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDefectNoteBody"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefectNote"

  # ── Bugs ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TestCase:
      type: object
      properties:
        id:
          type: integer
        useCaseId:
          type: integer
        caseNumber:
          type: integer
        title:
          type: string
        createdAt:
          type: string
      required: [id, useCaseId, caseNumber, title, createdAt]
...
    TestExecution:
      type: object
      properties:
        id:
          type: integer
        testCaseId:
          type: integer
        testRunId:
          type: integer
          nullable: true
        iterationNumber:
          type: integer
        testerName:
          type: string
        status:
          type: string
        executedAt:
          type: string
        stepResults:
          type: array
          items:
            $ref: "#/components/schemas/StepResult"
      required: [id, testCaseId, iterationNumber, testerName, status, executedAt]
...
    StepResult:
      type: object
      properties:
        id:
          type: integer
        executionId:
          type: integer
        stepId:
          type: integer
        actualResult:
          type: string
          nullable: true
        comments:
          type: string
          nullable: true
        passed:
          type: boolean
          nullable: true
        recordedAt:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
      required: [id, executionId, stepId, recordedAt, attachments]
...
    Attachment:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        field:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        createdAt:
          type: string
      required: [id, entityType, entityId, field, fileName, fileUrl, fileType, createdAt]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, username, name, role, isActive, createdAt]
...
    Defect:
      type: object
      properties:
        id:
          type: integer
        testRunId:
          type: integer
        testCaseId:
          type: integer
        executionId:
          type: integer
        testerNotes:
          type: string
          nullable: true
        status:
          type: string
        retestReason:
          type: string
          nullable: true
        acceptedByBusinessNote:
          type: string
          nullable: true
        rejectionLog:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        testCase:
          $ref: "#/components/schemas/TestCase"
          nullable: true
        execution:
          $ref: "#/components/schemas/TestExecution"
          nullable: true
        notes:
          type: array
          items:
            $ref: "#/components/schemas/DefectNote"
      required: [id, testRunId, testCaseId, executionId, status, createdAt, updatedAt]
    FlagRetestBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    BusinessAcceptBody:
      type: object
      properties:
        note:
          type: string
      required: [note]
    BusinessRejectBody:
      type: object
      properties:
        reason:
          type: string
    CreateDefectNoteBody:
      type: object
      properties:
        note:
          type: string
        discussionId:
          type: integer
          nullable: true
      required: [note]
    DefectNote:
      type: object
      properties:
        id:
          type: integer
        defectId:
          type: integer
        discussionId:
          type: integer
          nullable: true
        addedByUserId:
          type: integer
        note:
          type: string
        createdAt:
          type: string
          format: date-time
        addedByUser:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, defectId, addedByUserId, note, createdAt]
...